Yes, password removal is workng, but if you kill app during snackbar is shown, password won't be deleted. Need WorkManager

+++ app/src/main/java/lv/maros/secured/password/keeper/pages/passwords/PasswordsViewModel.kt
@@ -61,12 +61,14 @@ class PasswordsViewModel(
     }
 
     fun deletePasswords() {
-        viewModelScope.launch {
-            passwordsToDelete.forEach {
-                repository.deletePassword(it.id)
-            }
+        if (passwordsToDelete.isNotEmpty()) {
+            viewModelScope.launch {
+                passwordsToDelete.forEach {
+                    repository.deletePassword(it.id)
+                }
 
-            invalidateShowNoData()
+                invalidateShowNoData()
+            }
         }
     }
